requirements.txt
streamlit==1.31.0
google-generativeai==0.3.1
langchain==0.1.0
langchain-openai==0.0.5
faiss-cpu==1.7.4
pandas==2.1.4
numpy==1.26.0
python-dotenv==1.0.0
openai==1.6.1